Fix some integer casts incompatible with the new gtk-sys

+++ src/widgets/tree_path.rs
@@ -39,7 +39,7 @@ impl TreePath {
 
     #[cfg(gtk_3_12)]
     pub fn new_from_indicesv(indices: &mut [i32]) -> Option<TreePath> {
-        let tmp = unsafe { ffi::gtk_tree_path_new_from_indicesv(indices.as_mut_ptr(), indices.len() as ::libc::c_ulong) };
+        let tmp = unsafe { ffi::gtk_tree_path_new_from_indicesv(indices.as_mut_ptr(), indices.len() as ::libc::size_t) };
 
         if tmp.is_null() {
             None
